Struct aws_sdk_kendra::client::fluent_builders::CreateDataSource

pub struct CreateDataSource<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

Fluent builder constructing a request to CreateDataSource.

Creates a data source that you want to use with an Amazon Kendra index.

You specify a name, data source connector type and description for your data source. You also specify configuration information for the data source connector.

CreateDataSource is a synchronous operation. The operation returns 200 if the data source was successfully created. Otherwise, an exception is raised.

Implementations

impl<C, M, R> CreateDataSource<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub async fn send(
self
) -> Result<CreateDataSourceOutput, SdkError<CreateDataSourceError>> where
R::Policy: SmithyRetryPolicy<CreateDataSourceInputOperationOutputAlias, CreateDataSourceOutput, CreateDataSourceError, CreateDataSourceInputOperationRetryAlias>,

Sends the request and returns the response.

If an error occurs, an SdkError will be returned with additional details that can be matched against.

By default, any retryable failures will be retried twice. Retry behavior is configurable with the RetryConfig, which can be set when configuring the client.

pub fn name(self, inp: impl Into<String>) -> Self

A unique name for the data source. A data source name can't be changed without deleting and recreating the data source.

pub fn set_name(self, input: Option<String>) -> Self

A unique name for the data source. A data source name can't be changed without deleting and recreating the data source.

pub fn index_id(self, inp: impl Into<String>) -> Self

The identifier of the index that should be associated with this data source.

pub fn set_index_id(self, input: Option<String>) -> Self

The identifier of the index that should be associated with this data source.

pub fn type(self, inp: DataSourceType) -> Self

The type of repository that contains the data source.

pub fn set_type(self, input: Option<DataSourceType>) -> Self

The type of repository that contains the data source.

pub fn configuration(self, inp: DataSourceConfiguration) -> Self

The connector configuration information that is required to access the repository.

You can't specify the Configuration parameter when the Type parameter is set to CUSTOM. If you do, you receive a ValidationException exception.

The Configuration parameter is required for all other data sources.

pub fn set_configuration(self, input: Option<DataSourceConfiguration>) -> Self

The connector configuration information that is required to access the repository.

You can't specify the Configuration parameter when the Type parameter is set to CUSTOM. If you do, you receive a ValidationException exception.

The Configuration parameter is required for all other data sources.

pub fn description(self, inp: impl Into<String>) -> Self

A description for the data source.

pub fn set_description(self, input: Option<String>) -> Self

A description for the data source.

pub fn schedule(self, inp: impl Into<String>) -> Self

Sets the frequency that Amazon Kendra will check the documents in your repository and update the index. If you don't set a schedule Amazon Kendra will not periodically update the index. You can call the StartDataSourceSyncJob operation to update the index.

You can't specify the Schedule parameter when the Type parameter is set to CUSTOM. If you do, you receive a ValidationException exception.

pub fn set_schedule(self, input: Option<String>) -> Self

Sets the frequency that Amazon Kendra will check the documents in your repository and update the index. If you don't set a schedule Amazon Kendra will not periodically update the index. You can call the StartDataSourceSyncJob operation to update the index.

You can't specify the Schedule parameter when the Type parameter is set to CUSTOM. If you do, you receive a ValidationException exception.

pub fn role_arn(self, inp: impl Into<String>) -> Self

The Amazon Resource Name (ARN) of a role with permission to access the data source. For more information, see IAM Roles for Amazon Kendra.

You can't specify the RoleArn parameter when the Type parameter is set to CUSTOM. If you do, you receive a ValidationException exception.

The RoleArn parameter is required for all other data sources.

pub fn set_role_arn(self, input: Option<String>) -> Self

The Amazon Resource Name (ARN) of a role with permission to access the data source. For more information, see IAM Roles for Amazon Kendra.

You can't specify the RoleArn parameter when the Type parameter is set to CUSTOM. If you do, you receive a ValidationException exception.

The RoleArn parameter is required for all other data sources.

pub fn tags(self, inp: impl Into<Tag>) -> Self

Appends an item to Tags.

To override the contents of this collection use set_tags.

A list of key-value pairs that identify the data source. You can use the tags to identify and organize your resources and to control access to resources.

pub fn set_tags(self, input: Option<Vec<Tag>>) -> Self

A list of key-value pairs that identify the data source. You can use the tags to identify and organize your resources and to control access to resources.

pub fn client_token(self, inp: impl Into<String>) -> Self

A token that you provide to identify the request to create a data source. Multiple calls to the CreateDataSource operation with the same client token will create only one data source.

pub fn set_client_token(self, input: Option<String>) -> Self

A token that you provide to identify the request to create a data source. Multiple calls to the CreateDataSource operation with the same client token will create only one data source.

pub fn language_code(self, inp: impl Into<String>) -> Self

The code for a language. This allows you to support a language for all documents when creating the data source. English is supported by default. For more information on supported languages, including their codes, see Adding documents in languages other than English.

pub fn set_language_code(self, input: Option<String>) -> Self

The code for a language. This allows you to support a language for all documents when creating the data source. English is supported by default. For more information on supported languages, including their codes, see Adding documents in languages other than English.