Struct aws_sdk_kendra::client::fluent_builders::Query

pub struct Query<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

Implementations

impl<C, M, R> Query<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub async fn send(self) -> Result<QueryOutput, SdkError<QueryError>> where
R::Policy: SmithyRetryPolicy<QueryInputOperationOutputAlias, QueryOutput, QueryError, QueryInputOperationRetryAlias>,

pub fn index_id(self, inp: impl Into<String>) -> Self

The unique identifier of the index to search. The identifier is returned in the response from the CreateIndex operation.

pub fn set_index_id(self, input: Option<String>) -> Self

pub fn query_text(self, inp: impl Into<String>) -> Self

The text to search for.

pub fn set_query_text(self, input: Option<String>) -> Self

pub fn attribute_filter(self, inp: AttributeFilter) -> Self

Enables filtered searches based on document attributes. You can only provide one attribute filter; however, the AndAllFilters, NotFilter, and OrAllFilters parameters contain a list of other filters.

The AttributeFilter parameter enables you to create a set of filtering rules that a document must satisfy to be included in the query results.

pub fn set_attribute_filter(self, input: Option<AttributeFilter>) -> Self

pub fn facets(self, inp: impl Into<Facet>) -> Self

Appends an item to Facets.

To override the contents of this collection use set_facets.

An array of documents attributes. Amazon Kendra returns a count for each attribute key specified. You can use this information to help narrow the search for your user.

pub fn set_facets(self, input: Option<Vec<Facet>>) -> Self

pub fn requested_document_attributes(self, inp: impl Into<String>) -> Self

Appends an item to RequestedDocumentAttributes.

To override the contents of this collection use set_requested_document_attributes.

An array of document attributes to include in the response. No other document attributes are included in the response. By default all document attributes are included in the response.

pub fn set_requested_document_attributes(
self,
input: Option<Vec<String>>
) -> Self

pub fn query_result_type_filter(self, inp: QueryResultType) -> Self

Sets the type of query. Only results for the specified query type are returned.

pub fn set_query_result_type_filter(
self,
input: Option<QueryResultType>
) -> Self

pub fn document_relevance_override_configurations(
self,
inp: impl Into<DocumentRelevanceConfiguration>
) -> Self

Appends an item to DocumentRelevanceOverrideConfigurations.

To override the contents of this collection use set_document_relevance_override_configurations.

Overrides relevance tuning configurations of fields or attributes set at the index level.

If you use this API to override the relevance tuning configured at the index level, but there is no relevance tuning configured at the index level, then Amazon Kendra does not apply any relevance tuning.

If there is relevance tuning configured at the index level, but you do not use this API to override any relevance tuning in the index, then Amazon Kendra uses the relevance tuning that is configured at the index level.

If there is relevance tuning configured for fields at the index level, but you use this API to override only some of these fields, then for the fields you did not override, the importance is set to 1.

pub fn set_document_relevance_override_configurations(
self,
input: Option<Vec<DocumentRelevanceConfiguration>>
) -> Self

pub fn page_number(self, inp: i32) -> Self

Query results are returned in pages the size of the PageSize parameter. By default, Amazon Kendra returns the first page of results. Use this parameter to get result pages after the first one.

pub fn set_page_number(self, input: Option<i32>) -> Self

pub fn page_size(self, inp: i32) -> Self

Sets the number of results that are returned in each page of results. The default page size is 10. The maximum number of results returned is 100. If you ask for more than 100 results, only 100 are returned.

pub fn set_page_size(self, input: Option<i32>) -> Self

pub fn sorting_configuration(self, inp: SortingConfiguration) -> Self

Provides information that determines how the results of the query are sorted. You can set the field that Amazon Kendra should sort the results on, and specify whether the results should be sorted in ascending or descending order. In the case of ties in sorting the results, the results are sorted by relevance.

If you don't provide sorting configuration, the results are sorted by the relevance that Amazon Kendra determines for the result.

pub fn set_sorting_configuration(
self,
input: Option<SortingConfiguration>
) -> Self

pub fn user_context(self, inp: UserContext) -> Self

The user context token.

pub fn set_user_context(self, input: Option<UserContext>) -> Self

pub fn visitor_id(self, inp: impl Into<String>) -> Self

Provides an identifier for a specific user. The VisitorId should be a unique identifier, such as a GUID. Don't use personally identifiable information, such as the user's email address, as the VisitorId.

pub fn set_visitor_id(self, input: Option<String>) -> Self