Struct aws_sdk_kendra::input::create_index_input::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for CreateIndexInput

Implementations

impl Builder

pub fn name(self, input: impl Into<String>) -> Self

The name for the new index.

pub fn set_name(self, input: Option<String>) -> Self

pub fn edition(self, input: IndexEdition) -> Self

The Amazon Kendra edition to use for the index. Choose DEVELOPER_EDITION for indexes intended for development, testing, or proof of concept. Use ENTERPRISE_EDITION for your production databases. Once you set the edition for an index, it can't be changed.

The Edition parameter is optional. If you don't supply a value, the default is ENTERPRISE_EDITION.

For more information on quota limits for enterprise and developer editions, see Quotas.

pub fn set_edition(self, input: Option<IndexEdition>) -> Self

pub fn role_arn(self, input: impl Into<String>) -> Self

An Identity and Access Management(IAM) role that gives Amazon Kendra permissions to access your Amazon CloudWatch logs and metrics. This is also the role used when you use the BatchPutDocument operation to index documents from an Amazon S3 bucket.

pub fn set_role_arn(self, input: Option<String>) -> Self

pub fn server_side_encryption_configuration(
self,
input: ServerSideEncryptionConfiguration
) -> Self

The identifier of the KMScustomer managed key (CMK) to use to encrypt data indexed by Amazon Kendra. Amazon Kendra doesn't support asymmetric CMKs.

pub fn set_server_side_encryption_configuration(
self,
input: Option<ServerSideEncryptionConfiguration>
) -> Self

pub fn description(self, input: impl Into<String>) -> Self

A description for the index.

pub fn set_description(self, input: Option<String>) -> Self

pub fn client_token(self, input: impl Into<String>) -> Self

A token that you provide to identify the request to create an index. Multiple calls to the CreateIndex operation with the same client token will create only one index.

pub fn set_client_token(self, input: Option<String>) -> Self

pub fn tags(self, input: impl Into<Tag>) -> Self

pub fn set_tags(self, input: Option<Vec<Tag>>) -> Self

pub fn user_token_configurations(
self,
input: impl Into<UserTokenConfiguration>
) -> Self

pub fn set_user_token_configurations(
self,
input: Option<Vec<UserTokenConfiguration>>
) -> Self

pub fn user_context_policy(self, input: UserContextPolicy) -> Self

The user context policy.

ATTRIBUTE_FILTER

All indexed content is searchable and displayable for all users. If there is an access control list, it is ignored. You can filter on user and group attributes.

USER_TOKEN

Enables SSO and token-based user access control. All documents with no access control and all documents accessible to the user will be searchable and displayable.

pub fn set_user_context_policy(self, input: Option<UserContextPolicy>) -> Self

pub fn build(self) -> Result<CreateIndexInput, BuildError>

Consumes the builder and constructs a CreateIndexInput