Module aws_sdk_kendra::model

Modules

access_control_list_configuration

See AccessControlListConfiguration

acl_configuration

See AclConfiguration

additional_result_attribute

See AdditionalResultAttribute

additional_result_attribute_value

See AdditionalResultAttributeValue

attribute_filter

See AttributeFilter

authentication_configuration

See AuthenticationConfiguration

basic_authentication_configuration

See BasicAuthenticationConfiguration

batch_delete_document_response_failed_document

See BatchDeleteDocumentResponseFailedDocument

batch_get_document_status_response_error

See BatchGetDocumentStatusResponseError

batch_put_document_response_failed_document

See BatchPutDocumentResponseFailedDocument

capacity_units_configuration

See CapacityUnitsConfiguration

click_feedback

See ClickFeedback

column_configuration

See ColumnConfiguration

confluence_attachment_configuration

See ConfluenceAttachmentConfiguration

confluence_attachment_to_index_field_mapping

See ConfluenceAttachmentToIndexFieldMapping

confluence_blog_configuration

See ConfluenceBlogConfiguration

confluence_blog_to_index_field_mapping

See ConfluenceBlogToIndexFieldMapping

confluence_configuration

See ConfluenceConfiguration

confluence_page_configuration

See ConfluencePageConfiguration

confluence_page_to_index_field_mapping

See ConfluencePageToIndexFieldMapping

confluence_space_configuration

See ConfluenceSpaceConfiguration

confluence_space_to_index_field_mapping

See ConfluenceSpaceToIndexFieldMapping

connection_configuration

See ConnectionConfiguration

data_source_configuration

See DataSourceConfiguration

data_source_group

See DataSourceGroup

data_source_summary

See DataSourceSummary

data_source_sync_job

See DataSourceSyncJob

data_source_sync_job_metric_target

See DataSourceSyncJobMetricTarget

data_source_sync_job_metrics

See DataSourceSyncJobMetrics

data_source_to_index_field_mapping

See DataSourceToIndexFieldMapping

data_source_vpc_configuration

See DataSourceVpcConfiguration

database_configuration

See DatabaseConfiguration

document

See Document

document_attribute

See DocumentAttribute

document_attribute_value_count_pair

See DocumentAttributeValueCountPair

document_info

See DocumentInfo

document_metadata_configuration

See DocumentMetadataConfiguration

document_relevance_configuration

See DocumentRelevanceConfiguration

documents_metadata_configuration

See DocumentsMetadataConfiguration

facet

See Facet

facet_result

See FacetResult

faq_statistics

See FaqStatistics

faq_summary

See FaqSummary

google_drive_configuration

See GoogleDriveConfiguration

group_members

See GroupMembers

group_ordering_id_summary

See GroupOrderingIdSummary

group_summary

See GroupSummary

hierarchical_principal

See HierarchicalPrincipal

highlight

See Highlight

index_configuration_summary

See IndexConfigurationSummary

index_statistics

See IndexStatistics

json_token_type_configuration

See JsonTokenTypeConfiguration

jwt_token_type_configuration

See JwtTokenTypeConfiguration

member_group

See MemberGroup

member_user

See MemberUser

one_drive_configuration

See OneDriveConfiguration

one_drive_users

See OneDriveUsers

principal

See Principal

proxy_configuration

See ProxyConfiguration

query_result_item

See QueryResultItem

query_suggestions_block_list_summary

See QuerySuggestionsBlockListSummary

relevance

See Relevance

relevance_feedback

See RelevanceFeedback

s3_data_source_configuration

See S3DataSourceConfiguration

s3_path

See S3Path

salesforce_chatter_feed_configuration

See SalesforceChatterFeedConfiguration

salesforce_configuration

See SalesforceConfiguration

salesforce_custom_knowledge_article_type_configuration

See SalesforceCustomKnowledgeArticleTypeConfiguration

salesforce_knowledge_article_configuration

See SalesforceKnowledgeArticleConfiguration

salesforce_standard_knowledge_article_type_configuration

See SalesforceStandardKnowledgeArticleTypeConfiguration

salesforce_standard_object_attachment_configuration

See SalesforceStandardObjectAttachmentConfiguration

salesforce_standard_object_configuration

See SalesforceStandardObjectConfiguration

score_attributes

See ScoreAttributes

search

See Search

seed_url_configuration

See SeedUrlConfiguration

server_side_encryption_configuration

See ServerSideEncryptionConfiguration

service_now_configuration

See ServiceNowConfiguration

service_now_knowledge_article_configuration

See ServiceNowKnowledgeArticleConfiguration

service_now_service_catalog_configuration

See ServiceNowServiceCatalogConfiguration

share_point_configuration

See SharePointConfiguration

site_maps_configuration

See SiteMapsConfiguration

sorting_configuration

See SortingConfiguration

sql_configuration

See SqlConfiguration

status

See Status

suggestion

See Suggestion

suggestion_highlight

See SuggestionHighlight

suggestion_text_with_highlights

See SuggestionTextWithHighlights

suggestion_value

See SuggestionValue

tag

See Tag

text_document_statistics

See TextDocumentStatistics

text_with_highlights

See TextWithHighlights

thesaurus_summary

See ThesaurusSummary

time_range

See TimeRange

urls

See Urls

user_context

See UserContext

user_token_configuration

See UserTokenConfiguration

web_crawler_configuration

See WebCrawlerConfiguration

work_docs_configuration

See WorkDocsConfiguration

Structs

AccessControlListConfiguration

Access Control List files for the documents in a data source. For

AclConfiguration

Provides information about the column that should be used for

AdditionalResultAttribute

An attribute returned from an index query.

AdditionalResultAttributeValue

An attribute returned with a document from a search.

AttributeFilter

Provides filtering the query results based on document

AuthenticationConfiguration

Provides the configuration information to connect to websites that require

BasicAuthenticationConfiguration

Provides the configuration information to connect to websites that require

BatchDeleteDocumentResponseFailedDocument

Provides information about documents that could not be removed

BatchGetDocumentStatusResponseError

Provides a response when the status of a document could not be retrieved.

BatchPutDocumentResponseFailedDocument

Provides information about a document that could not be

CapacityUnitsConfiguration

Specifies capacity units configured for your enterprise edition index.

ClickFeedback

Gathers information about when a particular result was clicked by

ColumnConfiguration

Provides information about how Amazon Kendra should use the columns of

ConfluenceAttachmentConfiguration

Specifies the attachment settings for the Confluence data source.

ConfluenceAttachmentToIndexFieldMapping

Defines the mapping between a field in the Confluence data source

ConfluenceBlogConfiguration

Specifies the blog settings for the Confluence data source. Blogs

ConfluenceBlogToIndexFieldMapping

Defines the mapping between a blog field in the Confluence data

ConfluenceConfiguration

Provides configuration information for data sources that connect

ConfluencePageConfiguration

Specifies the page settings for the Confluence data source.

ConfluencePageToIndexFieldMapping

Defines the mapping between a field in the Confluence data source

ConfluenceSpaceConfiguration

Specifies the configuration for indexing Confluence spaces.

ConfluenceSpaceToIndexFieldMapping

Defines the mapping between a field in the Confluence data source

ConnectionConfiguration

Provides the information necessary to connect to a

DataSourceConfiguration

Configuration information for a Amazon Kendra data source.

DataSourceGroup

DataSourceSummary

Summary information for a Amazon Kendra data source. Returned in a call

DataSourceSyncJob

Provides information about a synchronization job.

DataSourceSyncJobMetricTarget

Maps a particular data source sync job to a particular data

DataSourceSyncJobMetrics

Maps a batch delete document request to a specific data source

DataSourceToIndexFieldMapping

Maps a column or attribute in the data source to an index field.

DataSourceVpcConfiguration

Provides information for connecting to an Amazon VPC.

DatabaseConfiguration

Provides the information necessary to connect a database to an

Document

A document in an index.

DocumentAttribute

A custom attribute value assigned to a document.

DocumentAttributeValueCountPair

Provides the count of documents that match a particular attribute

DocumentInfo

Identifies a document for which to retrieve status

DocumentMetadataConfiguration

Specifies the properties of a custom index field.

DocumentRelevanceConfiguration

Overrides the document relevance properties of a custom index field.

DocumentsMetadataConfiguration

Document metadata files that contain information such as the

Facet

Information about a document attribute

FacetResult

The facet values for the documents in the response.

FaqStatistics

Provides statistical information about the FAQ questions and

FaqSummary

Provides information about a frequently asked questions and answer

GoogleDriveConfiguration

Provides configuration information for data sources that connect

GroupMembers

A list of users or sub groups that belong to a group. Users and groups

GroupOrderingIdSummary

Information on the processing of PUT and DELETE actions

GroupSummary

HierarchicalPrincipal

Highlight

Provides information that you can use to highlight a search result

IndexConfigurationSummary

A summary of information about an index.

IndexStatistics

Provides information about the number of documents and the number

JsonTokenTypeConfiguration

Configuration information for the JSON token type.

JwtTokenTypeConfiguration

Configuration information for the JWT token type.

MemberGroup

The sub groups that belong to a group.

MemberUser

The users that belong to a group.

OneDriveConfiguration

Provides configuration information for data sources that connect

OneDriveUsers

User accounts whose documents should be indexed.

Principal

Provides user and group information for document access

ProxyConfiguration

Provides the configuration information for a web proxy to connect to

QueryResultItem

A single query result.

QuerySuggestionsBlockListSummary

Summary information on a query suggestions block list.

Relevance

Provides information for manually tuning the relevance of a field

RelevanceFeedback

Provides feedback on how relevant a document is to a search. Your

S3DataSourceConfiguration

Provides configuration information for a data source to index

S3Path

Information required to find a specific file in an Amazon S3

SalesforceChatterFeedConfiguration

Defines configuration for syncing a Salesforce chatter feed. The

SalesforceConfiguration

Provides configuration information for connecting to a Salesforce

SalesforceCustomKnowledgeArticleTypeConfiguration

Provides configuration information for indexing Salesforce custom

SalesforceKnowledgeArticleConfiguration

Specifies configuration information for the knowledge article

SalesforceStandardKnowledgeArticleTypeConfiguration

Provides configuration information for standard Salesforce

SalesforceStandardObjectAttachmentConfiguration

Provides configuration information for processing attachments to

SalesforceStandardObjectConfiguration

Specifies configuration information for indexing a single standard

ScoreAttributes

Provides a relative ranking that indicates how confident Amazon Kendra is

Search

Provides information about how a custom index field is used during

SeedUrlConfiguration

Provides the configuration information of the seed or starting point URLs to crawl.

ServerSideEncryptionConfiguration

Provides the identifier of the KMScustomer master key (CMK)

ServiceNowConfiguration

Provides configuration information required to connect to a

ServiceNowKnowledgeArticleConfiguration

Provides configuration information for crawling knowledge articles

ServiceNowServiceCatalogConfiguration

Provides configuration information for crawling service catalog

SharePointConfiguration

Provides configuration information for connecting to a Microsoft

SiteMapsConfiguration

Provides the configuration information of the sitemap URLs to crawl.

SortingConfiguration

Specifies the document attribute to use to sort the response to a

SqlConfiguration

Provides information that configures Amazon Kendra to use a SQL

Status

Provides information about the status of documents submitted for indexing.

Suggestion

A single query suggestion.

SuggestionHighlight

The text highlights for a single query suggestion.

SuggestionTextWithHighlights

Provides text and information about where to highlight the query suggestion text.

SuggestionValue

The SuggestionTextWithHighlights structure information.

Tag

A list of key/value pairs that identify an index, FAQ, or data source.

TextDocumentStatistics

Provides information about text documents indexed in an

TextWithHighlights

Provides text and information about where to highlight the

ThesaurusSummary

An array of summary information for one or more thesauruses.

TimeRange

Provides a range of time.

Urls

Provides the configuration information of the URLs to crawl.

UserContext

Provides information about the user context for a Amazon Kendra index.

UserTokenConfiguration

Provides configuration information for a token configuration.

WebCrawlerConfiguration

Provides the configuration information required for Amazon Kendra

WorkDocsConfiguration

Provides the configuration information to connect to Amazon WorkDocs

Enums

AdditionalResultAttributeValueType

ConfluenceAttachmentFieldName

ConfluenceBlogFieldName

ConfluencePageFieldName

ConfluenceSpaceFieldName

ConfluenceVersion

ContentType

DataSourceStatus

DataSourceSyncJobStatus

DataSourceType

DatabaseEngineType

DocumentAttributeValue

DocumentAttributeValueType

DocumentStatus

ErrorCode

FaqFileFormat

FaqStatus

HighlightType

IndexEdition

IndexStatus

KeyLocation

Mode

Order

PrincipalMappingStatus

PrincipalType

QueryIdentifiersEnclosingOption

QueryResultType

QuerySuggestionsBlockListStatus

QuerySuggestionsStatus

ReadAccessType

RelevanceType

SalesforceChatterFeedIncludeFilterType

SalesforceKnowledgeArticleState

SalesforceStandardObjectName

ScoreConfidence

Enumeration for query score confidence.

ServiceNowAuthenticationType

ServiceNowBuildVersionType

SharePointVersion

SortOrder

ThesaurusStatus

UserContextPolicy

WebCrawlerMode