Module aws_sdk_kendra::operation

Structs

BatchDeleteDocument

Removes one or more documents from an index. The documents must have

BatchGetDocumentStatus

Returns the indexing status for one or more documents submitted

BatchPutDocument

Adds one or more documents to an index.

ClearQuerySuggestions

Clears existing query suggestions from an index.

CreateDataSource

Creates a data source that you use to with an Amazon Kendra index.

CreateFaq

Creates an new set of frequently asked question (FAQ) questions and answers.

CreateIndex

Creates a new Amazon Kendra index. Index creation is an asynchronous

CreateQuerySuggestionsBlockList

Creates a block list to exlcude certain queries from suggestions.

CreateThesaurus

Creates a thesaurus for an index. The thesaurus

DeleteDataSource

Deletes an Amazon Kendra data source. An exception is not thrown if the

DeleteFaq

Removes an FAQ from an index.

DeleteIndex

Deletes an existing Amazon Kendra index. An exception is not thrown if

DeletePrincipalMapping

Deletes a group so that all users and sub groups that belong to the group can

DeleteQuerySuggestionsBlockList

Deletes a block list used for query suggestions for an index.

DeleteThesaurus

Deletes an existing Amazon Kendra thesaurus.

DescribeDataSource

Gets information about a Amazon Kendra data source.

DescribeFaq

Gets information about an FAQ list.

DescribeIndex

Describes an existing Amazon Kendra index

DescribePrincipalMapping

Describes the processing of PUT and DELETE actions

DescribeQuerySuggestionsBlockList

Describes a block list used for query suggestions for an index.

DescribeQuerySuggestionsConfig

Describes the settings of query suggestions for an index.

DescribeThesaurus

Describes an existing Amazon Kendra thesaurus.

GetQuerySuggestions

Fetches the queries that are suggested to your users.

ListDataSourceSyncJobs

Gets statistics about synchronizing Amazon Kendra with a data

ListDataSources

Lists the data sources that you have created.

ListFaqs

Gets a list of FAQ lists associated with an index.

ListGroupsOlderThanOrderingId

Provides a list of groups that are mapped to users before a

ListIndices

Lists the Amazon Kendra indexes that you have created.

ListQuerySuggestionsBlockLists

Lists the block lists used for query suggestions for an index.

ListTagsForResource

Gets a list of tags associated with a specified resource. Indexes,

ListThesauri

Lists the Amazon Kendra thesauri associated with an index.

PutPrincipalMapping

Maps users to their groups. You can also map sub groups to groups.

Query

Searches an active index. Use this API to search your documents

StartDataSourceSyncJob

Starts a synchronization job for a data source. If a synchronization

StopDataSourceSyncJob

Stops a running synchronization job. You can't stop a scheduled

SubmitFeedback

Enables you to provide feedback to Amazon Kendra to improve the

TagResource

Adds the specified tag to the specified index, FAQ, or data source

UntagResource

Removes a tag from an index, FAQ, or a data source.

UpdateDataSource

Updates an existing Amazon Kendra data source.

UpdateIndex

Updates an existing Amazon Kendra index.

UpdateQuerySuggestionsBlockList

Updates a block list used for query suggestions for an index.

UpdateQuerySuggestionsConfig

Updates the settings of query suggestions for an index.

UpdateThesaurus

Updates a thesaurus file associated with an index.